requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
scikit-learn>=1.4
pandas>=2.0
numpy>=1.26
werkzeug>=3.0
requests>=2.31
pdfminer.six>=20221105
authlib>=1.3
python-dotenv>=1.0
